% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{convert_book}
\alias{convert_book}
\title{Convert a quarto HTML book to O'Reilly's format}
\usage{
convert_book(src_path = ".", rendered_path = "_book", dest_path = "oreilly")
}
\arguments{
\item{src_path}{Path to quarto book source directory.}

\item{rendered_path}{Path to rendered quarto HTML book.}

\item{dest_path}{Path to save O'Reilly compatible book.}
}
\description{
Convert a quarto HTML book to O'Reilly's format
}
