% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minibook-package.R
\docType{package}
\name{htmlbook-package}
\alias{htmlbook}
\alias{htmlbook-package}
\title{htmlbook: Convert a Quarto book to O'Reilly's HTMLbook format}
\description{
This is a fairly quick and dirty package to post-process a Quarto html book into O'Reilly's htmlbook format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/htmlbook}
  \item Report bugs at \url{https://github.com/hadley/htmlbook/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
