% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minibook-package.R
\docType{package}
\name{minibook-package}
\alias{minibook}
\alias{minibook-package}
\title{minibook: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
