% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minibook-package.R
\docType{package}
\name{htmlbook-package}
\alias{htmlbook}
\alias{htmlbook-package}
\title{htmlbook: Convert a Quarto book to O'Reilly's html book format}
\description{
This is a fairly quick and dirty package to post-process a Quarto html book into O'Reilly's htmlbook format.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
